Rewrite discourse plugin, allow multiple consultation settings templates
