+++ plugins/antragsgruen_sites/Module.php
@@ -3,7 +3,6 @@
 namespace app\plugins\antragsgruen_sites;
 
 use app\models\db\Consultation;
-use app\models\layoutHooks\Hooks;
 use app\models\settings\Layout;
 use app\plugins\ModuleBase;
 
@@ -38,12 +37,7 @@ class Module extends ModuleBase
         }
     }
 
-    /**
-     * @param Layout $layoutSettings
-     * @param Consultation $consultation
-     * @return Hooks[]
-     */
-    public static function getForcedLayoutHooks($layoutSettings, $consultation)
+    public static function getForcedLayoutHooks(Layout $layoutSettings, ?Consultation $consultation)
     {
         return [
             new LayoutHooks($layoutSettings, $consultation)
