+++ plugins/discourse/messages/en/discourse.php
@@ -1,7 +1,12 @@
 <?php
 
 return [
-    'error_create' => 'Das Diskussionsthema im Grünen Forum konnte nicht angelegt werden. Bitte wende dich an beteiligung@gruene.de. Der Antrag selbst wurde angelegt.',
+    'error_create' => 'Das Diskussionsthema im Forum konnte nicht angelegt werden. Bitte wende dich an die Administrator*in.',
     'title_amend' => 'Beteiligungsgrün: Änderungsantrag zu %TITLE%, Zeile %LINE%',
     'title_motion' => 'Beteiligungsgrün: %TITLE%',
+
+    'sett_title' => 'Discourse-Integration',
+    'sett_discourse_cat' => 'Discourse Category-ID',
+
+    'goto_comments' => 'Go to comments',
 ];
