+++ plugins/frauenrat/Module.php
@@ -3,7 +3,6 @@
 namespace app\plugins\frauenrat;
 
 use app\models\db\Consultation;
-use app\models\layoutHooks\Hooks;
 use app\models\settings\Layout;
 use app\plugins\frauenrat\pdf\Frauenrat;
 use app\plugins\ModuleBase;
@@ -15,13 +14,7 @@ class Module extends ModuleBase
         parent::init();
     }
 
-    /**
-     * @param Layout $layoutSettings
-     * @param Consultation $consultation
-     *
-     * @return Hooks[]
-     */
-    public static function getForcedLayoutHooks($layoutSettings, $consultation)
+    public static function getForcedLayoutHooks(Layout $layoutSettings, ?Consultation $consultation)
     {
         return [
             new LayoutHooks($layoutSettings, $consultation)
