+++ plugins/gruen_gender/Module.php
@@ -3,18 +3,12 @@
 namespace app\plugins\gruen_gender;
 
 use app\models\db\Consultation;
-use app\models\layoutHooks\Hooks;
 use app\models\settings\Layout;
 use app\plugins\ModuleBase;
 
 class Module extends ModuleBase
 {
-    /**
-     * @param Layout $layoutSettings
-     * @param Consultation $consultation
-     * @return Hooks[]
-     */
-    public static function getForcedLayoutHooks($layoutSettings, $consultation)
+    public static function getForcedLayoutHooks(Layout $layoutSettings, ?Consultation $consultation)
     {
         return [
             new LayoutHooks($layoutSettings, $consultation)
