+++ plugins/member_petitions/SiteSpecificBehavior.php
@@ -2,7 +2,6 @@
 
 namespace app\plugins\member_petitions;
 
-use app\models\db\Consultation;
 use app\models\policies\IPolicy;
 use app\models\siteSpecificBehavior\DefaultBehavior;
 
@@ -35,18 +34,5 @@ class SiteSpecificBehavior extends DefaultBehavior
     {
         $controller = \Yii::$app->controller;
         return $controller->render('@app/plugins/member_petitions/views/index');
-    }
-
-    /**
-     * @param Consultation $consultation
-     * @return string
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
-     */
-    public static function getConsultationSettingsForm(Consultation $consultation)
-    {
-        return \Yii::$app->controller->renderPartial(
-            '@app/plugins/member_petitions/views/admin/consultation_settings',
-            ['consultation' => $consultation]
-        );
     }
 }
