+++ plugins/neos/Module.php
@@ -3,7 +3,6 @@
 namespace app\plugins\neos;
 
 use app\models\db\{Consultation, Site};
-use app\models\layoutHooks\Hooks;
 use app\models\settings\Layout;
 use app\models\siteSpecificBehavior\DefaultBehavior;
 use app\plugins\ModuleBase;
@@ -59,12 +58,7 @@ class Module extends ModuleBase
         return SiteSpecificBehavior::class;
     }
 
-    /**
-     * @param Layout $layoutSettings
-     * @param Consultation $consultation
-     * @return Hooks[]
-     */
-    public static function getForcedLayoutHooks($layoutSettings, $consultation)
+    public static function getForcedLayoutHooks(Layout $layoutSettings, ?Consultation $consultation)
     {
         return [
             new LayoutHooks($layoutSettings, $consultation)
